feat: create user authentication
