+++ src/app/service/account.service.ts
@@ -1,22 +1,19 @@
-import { HttpClient, HttpResponse } from "@angular/common/http"
+import { HttpClient } from "@angular/common/http"
 import { Injectable } from "@angular/core"
-import { BehaviorSubject, Observable, throwError, of } from "rxjs"
+import { BehaviorSubject, Observable } from "rxjs"
 import { map } from "rxjs/operators"
 import { User } from "../models/user"
 import { AlertService } from "./alert.service"
-import { UserService } from "./user.service"
-
+import { AUTH_TOKEN, KEY_TOKEN, BASE_URL, CURRENT_USER } from "src/app/constants/global"
 @Injectable({
   providedIn: "root",
 })
 export class AccountService {
-  private url: string = `http://localhost:3000`
-
   private currentUserSubject: BehaviorSubject<User>
   public currentUser: Observable<User>
 
   constructor(private http: HttpClient, public alertService: AlertService) {
-    this.currentUserSubject = new BehaviorSubject<User>(JSON.parse(localStorage.getItem("currentUser")))
+    this.currentUserSubject = new BehaviorSubject<User>(JSON.parse(localStorage.getItem(CURRENT_USER)))
     this.currentUser = this.currentUserSubject.asObservable()
   }
 
@@ -25,37 +22,18 @@ export class AccountService {
   }
 
   login({ email, password }) {
-    try {
-      return this.http.post<any>(`${this.url}/users/authenticate`, { email, password }).pipe(
-        map(user => {
-          window.localStorage.setItem("token", "meu-token")
-          this.currentUserSubject.next(user)
-          return user
-        })
-      )
-    } catch (error) {
-      console.log("✅ ~ error", error)
-    }
-
-    // const registeredUsers = allUsers.find(x => x.email === email && x.password === password)
-
-    // if (!registeredUsers) return throwError({ message: "Email ou senha incorretos." })
-
-    // const body = {
-    //   id: 0,
-    //   email: email,
-    //   token: "fake-jwt-token",
-    // }
-
-    // window.localStorage.setItem("currentUser", JSON.stringify(body))
-    // window.localStorage.setItem("token", "meu-token")
-    // this.currentUserSubject.next(body)
-
-    // return of(new HttpResponse({ status: 200, body }))
+    return this.http.post<any>(`${BASE_URL}/users/authenticate`, { email, password }).pipe(
+      map(user => {
+        const token = KEY_TOKEN
+        window.localStorage.setItem(AUTH_TOKEN, token)
+        this.currentUserSubject.next(user)
+        return user
+      })
+    )
   }
 
   logout() {
-    localStorage.removeItem("currentUser")
+    localStorage.removeItem(CURRENT_USER)
     this.currentUserSubject.next(null)
   }
 }
