+++ src/app/service/user.service.ts
@@ -1,22 +1,21 @@
 ﻿import { Injectable } from "@angular/core"
 import { HttpClient } from "@angular/common/http"
-
+import { BASE_URL } from "src/app/constants/global"
 import { User } from "../models/user"
 
 @Injectable({ providedIn: "root" })
 export class UserService {
-  private url: string = `http://localhost:3000`
   constructor(private http: HttpClient) {}
 
   getAll() {
-    return this.http.get<User[]>(`${this.url}/account`)
+    return this.http.get<User[]>(`${BASE_URL}/account`)
   }
 
   register(user: User) {
-    return this.http.post(`${this.url}/users/register`, user)
+    return this.http.post(`${BASE_URL}/users/register`, user)
   }
 
   delete(id: number) {
-    return this.http.delete(`${this.url}/users/${id}`)
+    return this.http.delete(`${BASE_URL}/users/${id}`)
   }
 }
