+++ src/app/shared/auth/auth.guard.ts
@@ -1,6 +1,7 @@
 import { Injectable } from "@angular/core"
 import { CanActivate, ActivatedRouteSnapshot, RouterStateSnapshot, Router } from "@angular/router"
 import { AccountService } from "../../service/account.service"
+import { AUTH_TOKEN } from "src/app/constants/global"
 
 @Injectable({
   providedIn: "root",
@@ -9,7 +10,7 @@ export class AuthGuard implements CanActivate {
   constructor(private router: Router, private accountService: AccountService) {}
 
   canActivate(next: ActivatedRouteSnapshot, state: RouterStateSnapshot): boolean {
-    const token = window.localStorage.getItem("token")
+    const token = window.localStorage.getItem(AUTH_TOKEN)
     const currentUser = this.accountService.currentUserValue
 
     if (token && currentUser) {
