+++ src/app/shared/error-msg/error-msg.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit, Input } from "@angular/core"
 import { FormControl } from "@angular/forms"
-import { FormValidations } from "../../helpers/form-validations"
+import { FormErrorMessages } from "../../helpers/form-error-messages"
 
 @Component({
   selector: "app-error-msg",
@@ -21,7 +21,7 @@ export class ErrorMsgComponent implements OnInit {
   get errorMessage() {
     for (const propertyName in this.control.errors) {
       if (this.control.errors.hasOwnProperty(propertyName) && this.control.touched) {
-        return FormValidations.getErrorMsg(this.label, propertyName, this.control.errors[propertyName])
+        return FormErrorMessages.getErrorMsg(this.label, propertyName, this.control.errors[propertyName])
       }
     }
